refactor: cleanup demo code, delete old components, use ReplicateUserFromTemplate final-form version only
