+++ src/data/repositories/UserD2ApiRepository.ts
@@ -528,7 +528,6 @@ export class UserD2ApiRepository implements UserRepository {
             searchOrganisationsUnits: this.getDomainOrgUnits(user.teiSearchOrganisationUnits),
             access: user.access,
             twoFA: userCredentials.twoFA,
-            // twoFactorEnabled: userCredentials.twoFactorEnabled,
             openId: userCredentials.openId,
             ldapId: userCredentials.ldapId,
             externalAuth: userCredentials.externalAuth,
@@ -577,7 +576,6 @@ export class UserD2ApiRepository implements UserRepository {
                 lastLogin: input.lastLogin?.toISOString() ?? "",
                 disabled: input.disabled,
                 twoFA: input.twoFA ?? false,
-                // twoFactorEnabled: input.twoFactorEnabled ?? false,
                 openId: input.openId ?? "",
                 ldapId: input.ldapId ?? "",
                 externalAuth: input.externalAuth ?? "",
@@ -650,7 +648,6 @@ const fields = {
         lastLogin: true,
         disabled: true,
         twoFA: true,
-        // twoFactorEnabled: true,
         openId: true,
         ldapId: true,
         externalAuth: true,
