+++ src/webapp/components/replicate/ReplicateUserFromTemplate.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useEffect } from "react";
+import _ from "lodash";
 
-import { ConfirmationDialog, useLoading, useSnackbar } from "@eyeseetea/d2-ui-components";
-import { DialogContent } from "@material-ui/core";
-import InfoDialog from "../../../legacy/components/InfoDialog";
+import { useLoading, useSnackbar } from "@eyeseetea/d2-ui-components";
+import { Dialog, DialogTitle, DialogActions, DialogContent, Button } from "@material-ui/core";
 import {
     InputField,
     composeValidators,
@@ -13,9 +13,12 @@ import {
     string,
     number,
 } from "@dhis2/ui";
+import { Form, FormSpy, Field } from "react-final-form";
+import { FormState } from "final-form";
 
 import i18n from "../../../locales";
 import { useAppContext } from "../../contexts/app-context";
+import { getFromTemplate } from "../../../utils/template";
 
 import { Id } from "../../../domain/entities/Ref";
 import { User, defaultUser } from "../../../domain/entities/User";
@@ -26,24 +29,16 @@ interface ReplicateUserFromTemplateProps {
     onRequestClose: () => void;
 }
 
-export const ReplicateUserFromTemplateFC: React.FC<ReplicateUserFromTemplateProps> = props => {
+export const ReplicateUserFromTemplate: React.FC<ReplicateUserFromTemplateProps> = props => {
     const { compositionRoot } = useAppContext();
     const { userToReplicateId, onRequestClose } = props;
 
     const [userToReplicate, setUserToReplicate] = React.useState<User>(defaultUser);
     const [existingUsernames, setExistingUsernames] = React.useState<string[]>([]);
     const [replicateTitle, setReplicateTitle] = React.useState<string>(i18n.t("Replicate User"));
-    const [replicateCount, setReplicateCount] = React.useState<string>("1");
-    const [usernameTemplate, setUsernameTemplate] = React.useState<string>("");
-    const [passwordTemplate, setPasswordTemplate] = React.useState<string>(`${UserLogic.DEFAULT_PASSWORD}_$index`);
-    const [hasValidationErrors, setValidationError] = React.useState<validationErrors>({
-        usersToCreate: false,
-        username: false,
-        password: false,
-    });
+    const [hasValidationErrors, setValidationError] = React.useState<boolean>(false);
     const [isUserLoaded, setIsUserLoaded] = React.useState(true);
     const [isMounted, setIsMounted] = React.useState(false);
-    const [infoDialog, setInfoDialog] = React.useState<{ response: string }>();
 
     const loading = useLoading();
     const snackbar = useSnackbar();
@@ -55,7 +50,7 @@ export const ReplicateUserFromTemplateFC: React.FC<ReplicateUserFromTemplateProp
         };
 
         loading.show(true);
-        setIsUserLoaded(true);
+        setIsUserLoaded(false);
 
         compositionRoot.users.get([userToReplicateId]).run(
             ([user]) => {
@@ -78,121 +73,133 @@ export const ReplicateUserFromTemplateFC: React.FC<ReplicateUserFromTemplateProp
                 handleUsersError(`Error loading user (${userToReplicateId}): ${error}`);
             }
         );
-
-        setIsUserLoaded(false);
+        setIsUserLoaded(true);
     }, [compositionRoot, loading, onRequestClose, snackbar, userToReplicateId]);
 
     useEffect(() => {
-        if (!isUserLoaded && userToReplicate.username) {
+        if (isUserLoaded && userToReplicate.username) {
             setReplicateTitle(
                 i18n.t("Replicate {{user}}", {
                     user: `${userToReplicate.name} (${userToReplicate.username})`,
                 })
             );
-            setUsernameTemplate(`${userToReplicate.username}_$index`);
             setIsMounted(true);
             loading.reset();
         }
     }, [isUserLoaded, loading, userToReplicate]);
 
-    const replicateUsers = useCallback(async () => {
-        loading.show(true, i18n.t("Replicating users"));
+    const replicateUsers = useCallback(
+        async ({
+            replicateCount,
+            usernameTemplate,
+            passwordTemplate,
+        }: {
+            replicateCount: string;
+            usernameTemplate: string;
+            passwordTemplate: string;
+        }) => {
+            loading.show(true, i18n.t("Replicating users"));
 
-        return compositionRoot.users
-            .replicateFromTemplate(userToReplicate, parseInt(replicateCount), usernameTemplate, passwordTemplate)
-            .run(
-                () => {
-                    loading.hide();
-                    onRequestClose();
-                    snackbar.success(
-                        i18n.t("User {{user}} replicated successfully {{n}} times", {
-                            user: userToReplicate.username,
-                            n: replicateCount,
-                        })
-                    );
-                },
-                error => {
-                    const errorMessage = i18n.t("Error replicating user {{user}}: {{message}}", {
-                        user: userToReplicate.username,
-                        message: error,
-                        nsSeparator: false,
-                    });
-                    loading.hide();
-                    snackbar.error(errorMessage);
-                }
-            );
-    }, [
-        compositionRoot.users,
-        loading,
-        snackbar,
-        onRequestClose,
-        userToReplicate,
-        replicateCount,
-        passwordTemplate,
-        usernameTemplate,
-    ]);
-
-    const CheckFormError = useCallback(
-        (label: string, error: boolean) => {
-            setValidationError({ ...hasValidationErrors, [label]: error });
+            return compositionRoot.users
+                .replicateFromTemplate(userToReplicate, parseInt(replicateCount), usernameTemplate, passwordTemplate)
+                .run(
+                    () => {
+                        loading.hide();
+                        onRequestClose();
+                        snackbar.success(
+                            i18n.t("User {{user}} replicated successfully {{n}} times", {
+                                user: userToReplicate.username,
+                                n: replicateCount,
+                            })
+                        );
+                    },
+                    error => {
+                        loading.hide();
+                        snackbar.error(
+                            i18n.t("Error replicating user {{user}}: {{message}}", {
+                                user: userToReplicate.username,
+                                message: error,
+                                nsSeparator: false,
+                            })
+                        );
+                    }
+                );
         },
-        [hasValidationErrors]
+        [compositionRoot.users, loading, snackbar, onRequestClose, userToReplicate]
     );
 
     return (
-        <ConfirmationDialog
-            isOpen={true}
-            title={replicateTitle}
-            maxWidth={"md"}
-            fullWidth={true}
-            onSave={replicateUsers}
-            saveText={i18n.t("Replicate")}
-            disableSave={Object.values(hasValidationErrors).some(validField => validField)}
-            onCancel={onRequestClose}
-        >
-            {infoDialog && (
-                <InfoDialog
-                    t={i18n.t}
-                    title={i18n.t("Replicate error")}
-                    onClose={() => setInfoDialog(undefined)}
-                    response={infoDialog.response}
-                />
-            )}
-
+        <>
             {isMounted && (
-                <DialogContent>
-                    <RenderInputField
-                        label={"Number of users to create"}
-                        value={replicateCount}
-                        setValue={setReplicateCount}
-                        validator={getValidators("usersToCreate").validation}
-                        setValidationError={CheckFormError}
-                    />
-                    <RenderInputField
-                        label={"Username template"}
-                        value={usernameTemplate}
-                        setValue={setUsernameTemplate}
-                        validator={getValidators("username", existingUsernames).validation}
-                        setValidationError={CheckFormError}
-                    />
-                    <RenderInputField
-                        label={"Password template"}
-                        value={passwordTemplate}
-                        setValue={setPasswordTemplate}
-                        validator={getValidators("password").validation}
-                        setValidationError={CheckFormError}
-                    />
-                </DialogContent>
+                <Dialog open maxWidth="lg" fullWidth>
+                    <DialogTitle>{replicateTitle}</DialogTitle>
+                    <DialogContent>
+                        <Form<FormValues>
+                            onSubmit={replicateUsers}
+                            initialValues={{
+                                replicateCount: "1",
+                                usernameTemplate: `${userToReplicate.username}_$index`,
+                                passwordTemplate: `${UserLogic.DEFAULT_PASSWORD}_$index`,
+                            }}
+                            validate={values => {
+                                const errors: {
+                                    replicateCount?: string;
+                                    usernameTemplate?: string;
+                                    passwordTemplate?: string;
+                                } = {};
+                                const replicateCountErrors = formValidator("usersToCreate").validation(
+                                    values.replicateCount
+                                );
+                                const usernameErrors = formValidator(
+                                    "username",
+                                    existingUsernames,
+                                    values.replicateCount
+                                ).validation(values.usernameTemplate);
+                                const passwordErrors = formValidator("password").validation(values.passwordTemplate);
+                                if (replicateCountErrors) errors.replicateCount = replicateCountErrors;
+                                if (usernameErrors) errors.usernameTemplate = usernameErrors;
+                                if (passwordErrors) errors.passwordTemplate = passwordErrors;
+                                return errors;
+                            }}
+                            autocomplete="off"
+                            render={({ handleSubmit }) => (
+                                <>
+                                    <FormSpy
+                                        onChange={(state: FormState<FormValues>) => {
+                                            requestAnimationFrame(() => {
+                                                setValidationError(!_.isEmpty(state.errors));
+                                            });
+                                        }}
+                                    />
+
+                                    <form id="replicate-form" onSubmit={handleSubmit}>
+                                        <DialogContent>
+                                            <RenderFormField name="replicateCount" label="Number of users to create" />
+                                            <RenderFormField name="usernameTemplate" label="Username template" />
+                                            <RenderFormField name="passwordTemplate" label="Password template" />
+                                        </DialogContent>
+                                    </form>
+                                </>
+                            )}
+                        />
+                    </DialogContent>
+
+                    <DialogActions>
+                        <Button onClick={onRequestClose}>{i18n.t("Cancel")}</Button>
+                        <Button disabled={hasValidationErrors} type="submit" form="replicate-form" color="primary">
+                            {i18n.t("Replicate")}
+                        </Button>
+                    </DialogActions>
+                </Dialog>
             )}
-        </ConfirmationDialog>
+        </>
     );
 };
 
-type FieldTypes = "usersToCreate" | "username" | "password";
-
-const getValidators = (
+const formValidator = (
     field: FieldTypes,
-    existingUsernames: string[] = []
+    existingUsernames: string[] = [],
+    count = "0"
 ): { validation: (...args: any[]) => string | undefined } => {
     switch (field) {
         case "usersToCreate": {
@@ -216,8 +223,16 @@ const getValidators = (
             return {
                 validation: (value: string) => {
                     if (!value) return i18n.t("Please provide a username");
+                    const countInt = parseInt(count);
+                    if (countInt > 1 && !value.includes("$index")) {
+                        return i18n.t("Username must contain $index");
+                    }
                     if (existingUsernames.includes(value)) {
                         return i18n.t("User already exists");
+                    }
+                    const usernameTemplate = _.times(countInt, index => getFromTemplate(value, index));
+                    if (_.intersection(usernameTemplate, existingUsernames).length > 0) {
+                        return i18n.t("Template will conflict with existing usernames");
                     } else {
                         const validators = composeValidators(
                             string,
@@ -255,32 +270,56 @@ const getValidators = (
     }
 };
 
-interface validationErrors {
-    usersToCreate: boolean;
-    username: boolean;
-    password: boolean;
-}
+const RenderFormField: React.FC<FormFieldProps> = ({ name, label }) => {
+    const helpText = label === "Password template" ? i18n.t("Remember to store the password") : "";
+    return (
+        <Field name={name}>
+            {({ input, meta }) => (
+                <RenderInputField
+                    label={label}
+                    value={input.value}
+                    setValue={input.onChange}
+                    error={meta.error}
+                    helpText={helpText}
+                />
+            )}
+        </Field>
+    );
+};
 
-interface InputFieldProps {
-    label: string;
-    value: string;
-    setValue: React.Dispatch<React.SetStateAction<string>>;
-    validator: (value: string) => string | undefined;
-    setValidationError: (label: string, error: boolean) => void;
-}
-
-const RenderInputField: React.FC<InputFieldProps> = ({ label, value, setValue, validator, setValidationError }) => {
-    const [error, setError] = React.useState<string | undefined>(undefined);
-
-    const handleChange = (data: { name?: string; value?: string }) => {
+const RenderInputField: React.FC<InputFieldProps> = ({ label, value, setValue, error, helpText }) => {
+    const handleChange = (data: { value?: string }) => {
         const newValue = data.value || "";
-        const validationError = validator(newValue);
-        setError(validationError);
-        validationError ? setValidationError(label, true) : setValidationError(label, false);
         setValue(newValue);
     };
 
-    return <InputField label={label} value={value} onChange={handleChange} validationText={error} error={!!error} />;
+    return (
+        <InputField
+            label={label}
+            value={value}
+            onChange={handleChange}
+            validationText={error}
+            error={!!error}
+            helpText={helpText}
+        />
+    );
 };
 
-export default ReplicateUserFromTemplateFC;
+type FormFieldLabels = "Number of users to create" | "Username template" | "Password template";
+type FieldTypes = "usersToCreate" | "username" | "password";
+type FormValues = { replicateCount: string; usernameTemplate: string; passwordTemplate: string };
+
+interface FormFieldProps {
+    name: string;
+    label: FormFieldLabels;
+}
+
+interface InputFieldProps {
+    label: FormFieldLabels;
+    value: string;
+    setValue: React.Dispatch<React.SetStateAction<string>>;
+    error: string | undefined;
+    helpText?: string;
+}
+
+export default ReplicateUserFromTemplate;
